requirements.txt
RPi.GPIO==0.7.1a4
click==8.0.3